+++ tests/testthat/test-DataMigrationManager.R
@@ -26,7 +26,7 @@ test_that("Migrations manager runs in package mode", {
   checkmate::expect_data_frame(manager$getStatus(), nrows = 2)
   expect_false(all(manager$getStatus()$executed))
   manager$executeMigrations()
-  manager$finalize()
+  manager$closeConnection()
 
   connection <- DatabaseConnector::connect(connectionDetails)
   on.exit(DatabaseConnector::disconnect(connection), add = TRUE)
@@ -43,7 +43,7 @@ test_that("Migrations manager runs in package mode", {
 
   checkmate::expect_data_frame(manager$getStatus(), nrows = 2)
   expect_true(all(manager$getStatus()$executed))
-  manager$finalize()
+  manager$closeConnection()
 })
 
 test_that("Migrations manager runs in folder mode", {
@@ -58,7 +58,7 @@ test_that("Migrations manager runs in folder mode", {
 
   expect_true(manager$check())
   manager$executeMigrations()
-  manager$finalize()
+  manager$closeConnection()
 
   connection <- DatabaseConnector::connect(connectionDetails)
   on.exit(DatabaseConnector::disconnect(connection), add = TRUE)
@@ -84,7 +84,7 @@ test_that("Add migration and execute", {
   expect_false(all(manager$getStatus()$executed))
   manager$executeMigrations()
   expect_true(all(manager$getStatus()$executed))
-  manager$finalize()
+  manager$closeConnection()
 })
 
 
@@ -101,7 +101,7 @@ test_that("Add invalid filename", {
     packageName = NULL
   )
   expect_false(manager$check())
-  manager$finalize()
+  manager$closeConnection()
 })
 
 test_that("Empty project works", {
@@ -115,5 +115,5 @@ test_that("Empty project works", {
   )
   checkmate::expect_data_frame(manager$getStatus(), nrows = 0)
   expect_true(manager$check())
-  manager$finalize()
+  manager$closeConnection()
 })