+++ tests/testthat/test-ResultExportManager.R
@@ -22,7 +22,7 @@ test_that("Test result export manager methods", {
 
   checkmate::expect_list(exportManager$getManifestList())
   exportManager$writeManifest()
-  expect_file_exists(file.path(export_dir, "manifest.json"))
+  checkmate::expect_file_exists(file.path(export_dir, "manifest.json"))
 })
 
 # Test exporting a data frame
@@ -70,13 +70,13 @@ test_that("exportDataFrame method exports data frame correctly", {
 
   expect_error(exportManager$exportDataFrame(df2, "table999", append = TRUE), "Table not found in specifications")
   exportManager$writeManifest()
-  expect_file_exists(file.path(export_dir, "manifest.json"))
+  checkmate::expect_file_exists(file.path(export_dir, "manifest.json"))
 })
 
 
 testDbExport <- function(connectionDetails, schema, n = 100) {
   ch <- ConnectionHandler$new(connectionDetails)
-  on.exit(ch$finalize())
+  on.exit(ch$closeConnection())
   # 1 million random rows
 
   table_spec <- dplyr::tibble(
@@ -98,8 +98,8 @@ testDbExport <- function(connectionDetails, schema, n = 100) {
   exportManager <- createResultExportManager(tableSpecification = table_spec, exportDir = exportDir, minCellCount = 5)
 
   transformFunc <- function(rows, pos, test) {
-    expect_data_frame(rows)
-    expect_integerish(pos)
+    checkmate::expect_data_frame(rows)
+    checkmate::expect_integerish(pos)
     expect_equal(test, 1234)
     return(rows)
   }