+++ NEWS.md
@@ -1,7 +1,12 @@
 # ResultModelManager 0.6.0
 
 Changes:
-1. Added optional wrapper functions for python based upload of csv files
+
+1. Added experimental wrapper functions for python based upload of csv files
+
+2. Breaking change - public use of finalize from R6 classes causes deprecation warning in use of R6 package.
+Package maintainers should replace this with class to `closeConnection` which is in use or has been added to all classes.
+Note - due to CRAN this package cannot release without making this change.
 
 Bug Fixes: